backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
networkx
python-dotenv
chromadb
sentence-transformers
pypdf
pytesseract
pdf2image